Dev daniel

+++ src/main/java/com/cibertec/security/repository/DepartamentoRepository.java
@@ -3,5 +3,5 @@ package com.cibertec.security.repository;
 import com.cibertec.security.entity.Departamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartamentoRepository extends JpaRepository<Departamento, Long> {
+public interface DepartamentoRepository extends JpaRepository<Departamento, Integer> {
 }
